+++ CHANGELOG.md
@@ -56,6 +56,7 @@ Changelog formatting (http://semver.org/):
 ### Changed
 
 - Switched from using bottom to top margin on .sources paragraph elements to allow spacing when no header is present
+- Clarified search page error results with suggestions
 
 ---